standardize program options with mkFormatterModule
A common complaint / observation is that treefmt-nix config tend to be a
mix between enabled programs, and treefmt settings to override
includes/excludes/priority. Some programs have those options available
directly, but not all of them.

In order to fix that, this commit introduces `mkFormatterModule`, a
function that generates a regular module structure based on a few
parameters. And that support the includes / excludes / priority out of
the box.

A few of the programs are already converted to that format. Ideally all
of them are getting converted soon.